+++ themes/Clean/views/user/account/_layout.php
@@ -6,7 +6,7 @@ use humhub\modules\user\widgets\ProfileHeader;
 $user = Yii::$app->user->identity;
 ?>
 
-<?php // Clean Theme: Add Profile header ?>
+<?php // Clean Theme: Add Profile header?>
 <?php if ($user): ?>
     <div id="account-profile-header" class="container">
         <div class="row">
@@ -17,9 +17,9 @@ $user = Yii::$app->user->identity;
     </div>
 <?php endif; ?>
 
-<?php require Yii::$app->getModule('user')->viewPath . '/account/_layout.php'; ?>
+<?php require Yii::getAlias('@user/views/account/_layout.php'); ?>
 
-<?php // Clean Theme: Replace "My account" button with "My profile" ?>
+<?php // Clean Theme: Replace "My account" button with "My profile"?>
 <script <?= Html::nonce() ?>>
     $(function () {
         $('#account-profile-header a.edit-account').attr('href', '<?= Yii::$app->user->identity->createUrl('/user/profile/home') ?>').text(<?= json_encode(Yii::t('base', 'My profile'), JSON_HEX_TAG) ?>);